% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_class.R
\name{object_class}
\alias{object_class}
\title{This is example function to demonstrate that devtools:check() can fail.}
\usage{
object_class(object)
}
\arguments{
\item{object}{R object}
}
\value{
The class of the object
}
\description{
This is example function to demonstrate that devtools:check() can fail.
}
\examples{
mat <- matrix(1:20, nrow=5, ncol=5)
object_class (mat)
}
